requirements.txt
beautifulsoup4
pandas
selenium
python-dotenv
requests
openpyxl